+++ src/anura/avss/__init__.py
@@ -3,7 +3,7 @@ from . import models as models
 from . import procedures as procedures
 from . import transport as transport
 from . import uuids as uuids
-from .client import AVSSClient, Report
+from .client import AVSSClient, ProgramTransferStats, Report
 from .exceptions import (
     AVSSBadArgumentError,
     AVSSConnectionError,
@@ -41,6 +41,7 @@ __all__ = [
     "CaptureReport",
     "HealthReport",
     "OpCode",
+    "ProgramTransferStats",
     "Report",
     "ReportType",
     "ResponseCode",